requirements.txt
streamlit
torch
transformers
pillow
pandas
fuzzywuzzy
python-Levenshtein